+++ site/news.md
@@ -4,6 +4,26 @@ title: News
 description: News and announcements for all things Bootstrap Table, including new releases.
 ---
 
+## Bootstrap Table 1.18.2
+
+<span class="post-date">23 Jan 2021</span>
+
+#### Core
+
+- **Update:** Fixed bootstrap5 cannot work bug.
+- **Update:** Fixed checkbox display bug when using `formatter`.
+- **Update:** Fixed search highlight bug.
+- **Update:** Updated `ru-RU` and `de-DE` locales.
+
+#### Extensions
+
+- **New(filter-control):** Added support for flat JSON.
+- **Update(cookie):** Fixed not deleted cookie bug when the sort was reset.
+- **Update(export):** Not export the detail view icon column.
+- **Update(filter-control):** Fixed not working when using `filterControlContainer`.
+- **Update(multiple-sort):** Fixed multiple-sort cannot work bug.
+- **Update(resizable):** Fixed resizable cannot work in modal.
+
 ## Bootstrap Table 1.18.1
 
 <span class="post-date">6 Dec 2020</span>